requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
ipykernel
langchain_groq
langchain_core
fastapi
pandas
openai
streamlit
html2text
validators
youtube_transcript_api
unstructured
pytube
numexpr